+++ netdef/Controllers/ModbusClientController.py
@@ -14,6 +14,85 @@ class ModbusClientController(BaseController.BaseController):
     """
     .. caution:: Development Status :: 4 - Beta
 
+    Read and write holding registers of a modbus device.
+
+    :param str name: The name is used i logfile and default.ini
+    :param Shared shared: reference to the global shared instance
+
+    Settings:
+
+    .. code-block:: ini
+
+        [ModbusClientController]
+
+        # connection
+        host = 127.0.0.1
+        port = 5020
+
+        # RUN_EXPRESSION is only sent if value has changed
+        oldnew_comparision = 1
+
+        # cooldown on connection error og write error
+        reconnect_timeout = 20
+
+        # Buffer or clear write requests recieved during cooldown
+        clear_writes_on_disconnect = 1
+
+        # Polling interval
+        poll_interval = 0.5
+
+    Sequence diagram:
+
+    .. seqdiag::
+
+        seqdiag app{
+            activation = none;
+            default_note_color = LemonChiffon;
+            span_height = 12;
+            edge_length = 200;
+
+            Queue [color=LemonChiffon];
+            Controller [label=ModbusClientController,color=LemonChiffon];
+            External [label="Modbus registers",color=LemonChiffon];
+
+            === Initialization ===
+            Queue -> Controller [label="APP_STATE, SETUP"]
+            === Setup ===
+            Queue -> Controller [label="ADD_SOURCE, source [n]"]
+            Queue -> Controller [label="APP_STATE, RUNNING"]
+            === Running ===
+            === Begin loop ===
+            Controller <- External [
+                label="Value change, register [n]",
+                leftnote="
+                    Update value of
+                    source [n]"
+            ]
+            Controller -> Queue [
+                label="RUN_EXPRESSION, source [n]",
+                note="
+                    Value change
+                    in source [n]"
+            ]
+            ... ...
+            Queue -> Controller [
+                label="
+                    WRITE_SOURCE,
+                    source [n], value, timestamp",
+                note="
+                    Update value of
+                    source [n]"
+            ]
+            Controller -> External [
+                label="update value, register [n]",
+                note="
+                    Value change
+                    in nodeid [n]"
+            ]
+            === End Loop ===
+        }
+
+
     """
     def __init__(self, name, shared):
         super().__init__(name, shared)
@@ -24,7 +103,7 @@ class ModbusClientController(BaseController.BaseController):
         self.clear_writes_on_disconnect = self.shared.config.config(self.name, "clear_writes_on_disconnect", 1)
         self.poll_interval = self.shared.config.config(self.name, "poll_interval", 0.5)
 
-        host = self.shared.config.config(self.name, "host", '0.0.0.0')
+        host = self.shared.config.config(self.name, "host", '127.0.0.1')
         port = self.shared.config.config(self.name, "port", 5020)
         self.client = ModbusClient(host, port=port)
 
@@ -66,15 +145,30 @@ class ModbusClientController(BaseController.BaseController):
         self.logger.info("Stopped")
 
     def safe_disconnect(self):
+        """
+        Close the tcp socket if it is connected
+        """
         try:
             self.client.close()
         except Exception as error:
             self.logger.warning("Cannot disconnect client: %s", error)
 
     def handle_add_source(self, incoming):
+        """
+        Add given source instance to internal source list
+
+        :param HoldingRegisterSource incoming: source instance
+        """
         self.add_source(incoming.key, incoming)
 
     def handle_write_source(self, incoming, value, source_time):
+        """
+        Write given value to the connected modbus device.
+
+        :param HoldingRegisterSource incoming: source instance
+        :param value: frozen value of instance
+        :param datetime.datetime source_time: value timestamp
+        """
         if hasattr(incoming, "unpack_unit_and_address"):
             slave_unit, register = incoming.unpack_unit_and_address()
 
@@ -98,6 +192,11 @@ class ModbusClientController(BaseController.BaseController):
                     )
 
     def poll_outgoing_item(self, item):
+        """
+        Poll given source for its value in the modbus device
+
+        :param HoldingRegisterSource item: source instance
+        """
         if hasattr(item, "unpack_unit_and_address"):
             slave_unit, register = item.unpack_unit_and_address()
             try: